source/api/requirements.txt
Django==3.2.4
djangorestframework==3.12.2
django-rest-enumfield==0.2.0